[feat][rename] 회원ID로 회원 조회기능 REST 방식으로 변경
