+++ user-service/src/main/java/org/omoknoone/omokhub/user/query/service/MemberService.java
@@ -1,35 +1,11 @@
 package org.omoknoone.omokhub.user.query.service;
 
 import org.omoknoone.omokhub.user.query.dto.MemberDTO;
-import org.omoknoone.omokhub.user.query.repository.MemberMapper;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 
-@Service
-public class MemberService {
+public interface MemberService {
+    MemberDTO searchMemberInfoByMemberId(String memberId);
 
-    private MemberMapper memberMapper;
+    MemberDTO searchMemberIdByNickname(String nickname);
 
-    public MemberService() {
-    }
-
-    @Autowired
-    public MemberService(MemberMapper memberMapper) {
-        this.memberMapper = memberMapper;
-    }
-
-    public MemberDTO searchMemberInfoByMemberId(String memberId) {
-
-        return memberMapper.selectMemberInfoByMemberId(memberId);
-    }
-
-    public MemberDTO searchMemberIdByNickname(String nickname) {
-
-        return memberMapper.selectMemberIdByNickname(nickname);
-    }
-
-    public Integer searchProfileCountByMemberId(String memberId) {
-
-        return memberMapper.selectProfileCountByMemberId(memberId);
-    }
+    Integer searchProfileCountByMemberId(String memberId);
 }
