+++ user-service/src/main/java/org/omoknoone/omokhub/user/query/service/PersonalLinkService.java
@@ -1,4 +1,4 @@
 package org.omoknoone.omokhub.user.query.service;
 
-public class PersonalLinkService {
+public interface PersonalLinkService {
 }
