+++ user-service/src/main/java/org/omoknoone/omokhub/user/query/service/PersonalLinkServiceImpl.java
@@ -1,0 +1,7 @@
+package org.omoknoone.omokhub.user.query.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class PersonalLinkServiceImpl implements PersonalLinkService {
+}
