+++ user-service/src/main/java/org/omoknoone/omokhub/user/query/service/ProfileService.java
@@ -1,4 +1,4 @@
 package org.omoknoone.omokhub.user.query.service;
 
-public class ProfileService {
+public interface ProfileService {
 }
