+++ user-service/src/main/resources/org/omoknoone/omokhub/user/query/repository/MemberMapper.xml
@@ -26,7 +26,7 @@
             ADDRESS ,
             BIRTHDAY
         FROM
-            MEMBER
+            member
         WHERE
             IS_WITHDRAW = 0
             AND MEMBER_ID = #{memberId}
